Extra fixes for the editor.
Refactored the regiondrag methods to be easier to understand.
Renamed some controls to be properly namespaced.
Fixed not being able to erase coins by assigning a block's value instead of simply checking if it's a coin, this means you won't be able to droppertool coins, however.
Introduced shortcuts for the tiles tabs as ctrl+1-7 by default.