Added animationtarget entity & animation event so that animations can cause linked events.
Added keyboard shortcuts to change levels within a world (kp4 and kp6).
Made the axe a dedicated entity. Still some stray code cleanup to do.
Changed the way mario handles redcoin/axe, if a collided entity has a "collect" method then mario will call it on contact with any face.
Added entity to play a sound effect from the soundlist, works like musicentity.
Added parameters to textentity to allow the text's starting position to be anywhere within the block it starts in.
Added peach/toad to the base SMB tileset because that's what they are. Tiles.