Massive overhaul of the way the editor uses tools.
Reverted default gamemode type because it was breaking things.
Fixed the linker tool and region tool. To save your changes to the region editor, click outside the boundary of it and then rightclick. It's really janky until I push certain bits of code into the parent.
Made the gui system for buttons optionally send a signal upon button release, decreasing the necessity of the editorignore* flags.
Light draw tool will probably cause a giant mess if used, don't do that.
Player-moving with rightclick temporarily disabled because the tool is incomplete.
Made tools directory that the editor autoloads tools from.
Added table.print as a utility function. Added nop as a global function that does nothing.