Made tile selection screen have a dropdown to choose from, keyboard shortcuts are updated accordingly.
Changed draw order for the black background so that tooltips weren't covered by the main tabs.
Added calibration method to entitylistitem, which is necessary to get correct highlights out of it due to "offy" and being on a scrollable index.
Prevented entitytooltip from attempting to draw with 0 alpha, offset the tooltip to appear at the corner of the cursor when possible.
Map tabs now have a less-sensitive vertical scrollbar.
Dropper updated to change to correct tile tab.
Added utility function table.find for finding an index in a table where a given key exists.
Removed "thearea" global because it was scoped to the corresponding gui elements.