Implemented dynamic audio positions. This means things will dopple and sound funny until we fine-tune the clamps.
If the audio sounds weird, it's because we don't have a way of globally updating the position of a sound each frame, so the supplied velocity is throwing it off.
Flagged all sounds that do not use a position as being allowed.
Enemies and item spawns are probably playing duplicates of themselves and are not responsible for their own noises but we can deal with that later.

Made music relative to the listener, set to the listener's position, for the sake of making it global.
Made portals responsible for their fizzle sound.
Made the portal entering sound local to the player.
Made fireballs responsible for their own sounds.
Made the vine responsible for the sound of itself spawning.