Converted box into a basedent, which gives footing for other objects to be carried.
Reduced debug prints in fireball.
Remapped basedents mixins to be global.
Renamed addOutput in baseent to addoutput for compatability with older entities.
Allowed basedents to be spawned into the map.
Reduced the size of the box graphic to allow basedent to demand a size.
Made player use the emancipate method of its held item when emancipated.
Renamed player method "pickupbox" to "pick_up".
Renamed player method "dropbox" to "drop_held".
Added a debug print for special physics case with boxes and players, let me know if you see it in the wild.