Overhauled enemies to have a system of taking damage from specific type sources.
Enemies have a generic do_damage method that attempts to perform a callback for a given damage type.
Enemies take damage directly via hurt method, which determines if it should apply the damage by using method evaluate_damagetype.
Gave enemies flag "gamemodeexclusive", which then checks "exclusivegamemode" to see if the entity should only exist in a given gamemode and marks it for deletion if it's not currently that gamemode.

Localized coins, lives, and score to the player object. There are still stray references to mariolives, score, and global coin counters.
Gave the player a murder method to act as a buffer to insert a message in the killfeed before signaling "die".
Fixed the HUD to reference player score and coincount.
Gave the koopa a damagetype of shell, this needs to be fixed so that it's only damage of type shell when in their shell.
Commented out mysterious negative score being given to players by the coinblockanimation.
Changed the domain of the addpoints/addcoins animation events to require a target player.
Made mixin for map-placeable basedentities.
Made basedentities compatible with new damagetype system, corrected existing ones. This still requires making more generic accessors.
Added killfeed system to track who kills what and why.
Added combo and score enums so that each damage type can be combo'd. This temporarily breaks all instances of player.combo and the conditions that reset them.