Killfeed upgrade and player bugfixes.
Killfeed now traces humiliations for when the victim and the attacker are the same.
Killfeed treats a nil attacker or victim as "world".
Killfeed highlights gold if the attacker or victim is a local player.
Player gets new traceinfluence to walk an entity to try and find a player responsible for an action.