Added a demo to illustrate a possible pixel transform function that maps colors to pixels without the use of a shader.
This comes at the expense of loading an instance of the data into memory and then converting it to a different type, which preserves a reference to the original image.
Since there is no way to clone image data, we end up with 8 or 9 image instances for only 4 variants of an image.