Made doing damage it's own behavior. (CanDamage)
Added damage callback for physics.
Made carried objects inherit parent speed so they don't lag behind.
CanInfluence and CanDamage should probably share the same properties because they depend on one another.
Made it so that enemies that are stompable are stomped by boxes. This makes the property "killedbyboxes" no longer necessary, as it's a logical default behavior. We can later implement an exception property if we determine we don't like this.