Refactored baseentity to no longer pass origclass and classname because those are available through middleclass' .class and .class.name vars respectively.
Made it so that baseentity reads the r table (from when the map passes data in) in a less-destructive way.
Basedentities can now load in data from the r table using the EDITOR_RCM static and a special attribute "name".

Gave the box an alternate graphic by alesan99 with some modifications by myself as an example of using its rightclick menu.