Repaired portaling and scrolling to follow the player.
Scrolling is stubbed due to the editor being unloaded and never setting the minimapdragging global. Take this as a lesson for expecting explicit false values from flags in lua.
The code is still scoped strangely (attempting to move things such that they're managed by either the portal object or the portalgun).
Returned ylookmodifier,autoscroll,everyonedead,noupdate to the global scope so the game can reset upon player death.
Rescoped portal gun drawing to be managed by the portalgun.
Rescoped getportalposition and traceline to the portalgun.
Traceline could be rescoped to the weapon object for generic finding.
Added dummy global var collections engine and gamesets.
Portal object now requires a parent if being created.