Borrowed the parse method from cupid to allow the console to not choke and die on any confusing inputs.
Merged console.lua into debug_bar.
Rescoped the console bar to be self contained, but didn't make it a class yet.

Changed the add method to use os.clock() because love.timer.getTime() is thrown out of whack by the love.run method.
Added helper setMouseContext to easily set the mouse cursor.
Added function playerDebug to be overloaded in the console whenever a debug button is pressed.

Changed the timing of logoPresent and the resources allocated for it.
Added helpers printwithfont and printfwithfont to wrap love.graphics.print* and use a particular font.
Added global resources table to minimize the amount of stuff being allocated.
Set the default resolution in conf.lua until a way to initialize the window in its target resolution crops up.

Gave loveframes textinput object a method to set the lines, because it's sort of weird to have a getter without a setter. Also the logic to work around it is crazy difficult.

Added a pixelfont example because drawing each individual letter using quads is actually silly when we're not doing transforms on each character like a rareware title.