Included the PROBE library to replace the broken stamp implementation, cloned grapher to provide a private instance dedicated to stacked-graphs.
Allowed grapher to pass arguments to an update function.
Added a pointUpdate method to grapher to allow updating without spreading it over the delta of time.
Added a flag (dlabel) in grapher to disallow label drawing.
Added options (cscalemin, cscalemax) in grapher to impact the scaling of the graph heights. (I don't know why I did this.)
Added graphSort and graphsUnmanaged to grapher to allow sorting the graphs internally by a certain feature (minimum value).
Removed redundant TLbind update.
Removed updating the window title per-frame to reduce impact on FPS (that it was ironically meant to measure).
Added a keypressed exception handler because the way PROBE works, if it was handled by an input manager it would start a cycle prematurely.
Fixed a bug when hitting down in the console bar causing a crash.
Renamed the love.run function to temporarily use the default handler.
Added hooks to lurker for pre and post scan functions.
Added a key for lurker (s) to force a scan on the crash screen.
Pushed a handful of the debug globals into the game global container.
Gave math.scale default values.